pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexus-chat"
version = "0.1.0"
description = "A modular, agentic chatbot with RAG support and multi-LLM provider integration"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Nexus Chat Contributors"},
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=13.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "aiofiles>=24.1.0",
    "anthropic>=0.34.0",
    "openai>=1.45.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "PyPDF2>=3.0.0",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.6.0",
]

[project.scripts]
nexus-chat = "backend.main:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
